refactor endpoints for external access
