+++ app/aimodels/bertopic/routers/summarization.py
@@ -14,6 +14,65 @@ router = APIRouter(
     prefix=""
 )
 
+@router.get(
+    "/model/{id}/visualize_topic_clusters",
+    response_class=HTMLResponse,
+    summary="Retrieve a BERTopic document cluster visualization",
+    response_description="Retrieved a BERTopic document cluster visualization")
+async def visualize_topic_clusters(id: UUID4, db: Session = Depends(get_db)):
+    """
+    Retrieve a BERTopic document cluster visualization
+
+    - **trained_model_id**: Required.  Trained BERTopic model ID.
+    """
+
+    model_obj = crud.bertopic_trained.get(db, id)
+    if not model_obj:
+        raise HTTPException(
+            status_code=422, detail=f"BERTopic trained model not found")
+
+    return model_obj.topic_cluster_visualization
+
+
+@router.get(
+    "/model/{id}/visualize_topic_words",
+    response_class=HTMLResponse,
+    summary="Retrieve a BERTopic word probability visualization",
+    response_description="Retrieved a BERTopic word probability visualization")
+async def visualize_topic_words(id: UUID4, db: Session = Depends(get_db)):
+    """
+    Retrieve a BERTopic word probability visualization
+
+    - **trained_model_id**: Required.  Trained BERTopic model ID.
+    """
+
+    model_obj = crud.bertopic_trained.get(db, id)
+    if not model_obj:
+        raise HTTPException(
+            status_code=422, detail=f"BERTopic trained model not found")
+
+    return model_obj.topic_word_visualization
+
+
+@router.get(
+    "/model/{id}/topics",
+    response_model=Union[list[TopicSummary], HTTPValidationError],
+    summary="Retrieve topics for a trained BERTopic model",
+    response_description="Retrieved topics for a trained BERTopic model")
+async def get_model_topics(id: UUID4, db: Session = Depends(get_db)) -> (
+    Union[list[TopicSummary], HTTPValidationError]
+):
+    """
+    Retrieve topics for a trained BERTopic model
+
+    - **trained_model_id**: Required.  Trained BERTopic model ID.
+    """
+
+    if not crud.bertopic_trained.get(db, id):
+        raise HTTPException(
+            status_code=422, detail=f"BERTopic trained model not found")
+
+    return crud.topic_summary.get_by_model_id(db, model_id=id)
 
 @router.get(
     "/topic/{id}",
