+++ app/aimodels/gpt4all/prefix_router.py
@@ -3,7 +3,7 @@ from .routers.completions import router as gpt4all_completions_router
 from .routers.pretrained import router as gpt4all_pretrained_router
 
 router = APIRouter(
-    prefix="/gpt4all", tags=["GPT4All"]
+    prefix="/gpt4all", tags=["GPT4All - Experimental"]
 )
 
 router.include_router(gpt4all_completions_router)
