+++ app/aimodels/lstm_stress_classifier/prefix_router.py
@@ -2,7 +2,7 @@ from fastapi import APIRouter
 from .routers.upload import router as upload_router
 
 router = APIRouter(
-    prefix="/lstmstressclassifier", tags=["LSTM Stress Classifier"]
+    prefix="/lstmstressclassifier", tags=["LSTM Stress Classifier - Experimental"]
 )
 
 router.include_router(upload_router)
