+++ app/main.py
@@ -7,10 +7,8 @@ from fastapi_versioning import VersionedFastAPI
 from .core.config import settings, OriginationEnum, set_acronym_dictionary
 from .core.logging import logger, LogConfig
 from logging.config import dictConfig
+from .experimental_features_router import router as experimental_router
 from .aimodels.router import router as aimodels_router
-from .sentiments.router import router as sentiments_router
-from .mattermost.router import router as mattermost_router
-from .chat_search.routers.retrieval import router as chat_queries_router
 from .dependencies import httpx_client
 
 dictConfig(LogConfig().dict())
@@ -39,9 +37,7 @@ origins = [
 ]
 
 app.include_router(aimodels_router)
-app.include_router(sentiments_router)
-app.include_router(mattermost_router)
-app.include_router(chat_queries_router)
+app.include_router(experimental_router)
 
 # set originated_from for standard app usage
 @app.get('/originated_from_app/')
